+++ bundles/org.openhab.binding.yioremote/src/main/java/org/openhab/binding/yioremote/internal/DTOs/IRReceiverMessage.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.yioremote.internal.DTOs;
+package org.openhab.binding.yioremote.internal.dtos;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
