+++ bundles/org.openhab.binding.yioremote/src/main/java/org/openhab/binding/yioremote/internal/YIOremoteDockActions.java
@@ -48,6 +48,8 @@ public class YIOremoteDockActions implements ThingActions {
             case AUTHENTICATION_COMPLETE:
                 dockHandler.sendIRCode(irCode);
                 break;
+            default:
+                break;
         }
     }
